pyproject.toml
[tool.poetry]
name = "dgs_test_2"
version = "0.1.0"
description = ""
authors = ["vildan-valeev <vildan.maydanovich.valeev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
requests = "^2.27.1"
redis = "^4.2.2"
psycopg2-binary = "^2.9.3"
python-decouple = "^3.6"
django-stubs-ext = "^0.4.0"
django-split-settings = "^1.1.0"
gunicorn = "^20.1.0"
django-dramatiq = "^0.10.0"
drf-yasg = "^1.20.0"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
django-filter = "^21.1"
flake8 = "^4.0.1"
pylint = "^2.13.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
